Centralize format_file_size utility function
Moved the format_file_size function to app/utils/generic.py for reuse across modules. Updated imports and removed duplicate implementations in file_search_controller.py, mod_sorting.py, mod_info_panel.py, and mods_panel.py to use the centralized utility.
