+++ app/sort/mod_sorting.py
@@ -161,18 +161,6 @@ def get_mod_metadata(uuid: str) -> Optional[ModMetadata]:
     return MetadataManager.instance().internal_local_metadata.get(uuid)
 
 
-def format_file_size(size_in_bytes: int) -> str:
-    """Format bytes to a human-readable string."""
-    if size_in_bytes < 1024:
-        return f"{size_in_bytes} B"
-    elif size_in_bytes < 1024 * 1024:
-        return f"{size_in_bytes / 1024:.1f} KB"
-    elif size_in_bytes < 1024 * 1024 * 1024:
-        return f"{size_in_bytes / (1024 * 1024):.1f} MB"
-    else:
-        return f"{size_in_bytes / (1024 * 1024 * 1024):.2f} GB"
-
-
 class ModsPanelSortKey(Enum):
     """
     Enum class representing different sorting keys for mods.
